pyproject.toml
[tool.poetry]
name = "hotel-booking"
version = "0.1.0"
description = "A simple hotel room and booking management service"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
pydantic-settings = "^2.0"
ruff = "^0.3"
pytest = "^8.0"
pytest-django = "^4.8"
dj-database-url = "^2.1"

[tool.poetry.group.dev.dependencies]  # Обновленный синтаксис
pre-commit = "^3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "hotel_booking.settings"
python_files = ["tests/*.py"]